Implement response speed and code cleanup

+++ src/main/kotlin/com/fcfb/discord/refbot/model/fcfb/game/Play.kt
@@ -3,7 +3,6 @@ package com.fcfb.discord.refbot.model.fcfb.game
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
-import java.math.BigInteger
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
 data class Play(
@@ -39,6 +38,6 @@ data class Play(
     @JsonProperty("home_timeouts") val homeTimeouts: Int,
     @JsonProperty("away_timeouts") val awayTimeouts: Int,
     @JsonProperty("play_finished") val playFinished: Boolean,
-    @JsonProperty("offensive_response_speed") val offensiveResponseSpeed: BigInteger?,
-    @JsonProperty("defensive_response_speed") val defensiveResponseSpeed: BigInteger?,
+    @JsonProperty("offensive_response_speed") val offensiveResponseSpeed: Long?,
+    @JsonProperty("defensive_response_speed") val defensiveResponseSpeed: Long?,
 )
